Export defaultReporter
I'm writing a checker using scss-lint, and I'd like to be able to use a custom reporter but *also* use the default reporter.

I'm using `gulp-notify` to create a desktop notification, but I don't want to reinvent the wheel and write a new console notification.

I want to define a custom reporter like this:

    var scsslint = require('gulp-scss-lint');
    
    if (!file.scsslint.success) {
      notifier.notify({ options});
      scsslint.defaultReporter(file);
    }

I think this line is all I need to make that possible. Seem reasonable?